3D Cost Calculation
This project is a desktop application developed to facilitate the processes of calculating the cost of your 3D model. The application allows users to quickly and effectively calculate the costs of 3D models.

Programming Languages Used:
Java
Python
Instructions:
The program for our application is the call_java.exe file located in the dist folder within the scripts directory.

If the command prompt file opens and closes immediately when you run the application, you need to do the following:
Copy the name of the "src" folder from the line src_path = "src" in the Python file named call_java.py and replace it with "src". For example, src_path = "C://Desktop//3DCostCalculation//src".